removed old react components + fixed elapsedTime
