+++ fbw-a32nx/src/systems/instruments/src/Clock/Clock.tsx
