+++ fbw-a32nx/src/systems/instruments/src/Clock/Components/Chrono.tsx
@@ -1,56 +1,93 @@
-import React, { useEffect, useState } from 'react';
-import { useInteractionEvent } from '../../Common/hooks';
-import { useSimVar } from '../../Common/simVars';
-import { debouncedTimeDelta } from '../../Common';
+import { ComponentProps, DisplayComponent, EventBus, FSComponent, HEvent, Subject, VNode } from 'msfssdk';
+import { debouncedTimeDelta } from '../shared/Utils';
+import { ClockSimvars } from '../shared/ClockSimvarPublisher';
 
-const getDisplayString = (seconds: number | null, running: boolean) : string => (seconds == null ? ''
-    : `${Math.floor(Math.min(seconds, 5999) / 60).toString().padStart(2, '0')}${running ? ':' : ' '}${(Math.floor(Math.min(seconds, 5999) % 60)).toString().padStart(2, '0')}`);
+interface ChronoProps extends ComponentProps {
+    bus: EventBus;
+}
 
-export const Chrono = () => {
-    const [ltsTest] = useSimVar('L:A32NX_OVHD_INTLT_ANN', 'bool', 250);
-    const [dcEssIsPowered] = useSimVar('L:A32NX_ELEC_DC_ESS_BUS_IS_POWERED', 'bool', 250);
-    const [absTime] = useSimVar('E:ABSOLUTE TIME', 'Seconds', 200);
-    const [prevTime, setPrevTime] = useState(absTime);
+const getDisplayString = (seconds: number | null, running: boolean, ltsTest: boolean) : string => {
+    if (ltsTest) {
+        return '88:88';
+    }
 
-    const [elapsedTime, setElapsedTime] = useState<null | number>(null);
-    const [, setElapsedSimvarTime] = useSimVar('L:A32NX_CHRONO_ELAPSED_TIME', 'number', 250);
-    const [running, setRunning] = useState(false);
-
-    useEffect(() => {
-        if (running) {
-            const newElapsedTime = (elapsedTime || 0) + debouncedTimeDelta(absTime, prevTime);
-            setElapsedTime(newElapsedTime);
-            setElapsedSimvarTime(newElapsedTime);
-        }
-        setPrevTime(absTime);
-    }, [absTime]);
-
-    useEffect(() => {
-        setElapsedSimvarTime(-1);
-    }, []);
-
-    useInteractionEvent('A32NX_CHRONO_TOGGLE', () => {
-        if (dcEssIsPowered) {
-            if (running) {
-                setRunning(false);
-            } else {
-                setRunning(true);
-            }
-        }
-    });
-    useInteractionEvent('A32NX_CHRONO_RST', () => {
-        if (dcEssIsPowered) {
-            if (running) {
-                setElapsedTime(0);
-                setElapsedSimvarTime(0);
-            } else {
-                setElapsedTime(null);
-                setElapsedSimvarTime(-1); // Simvar is not nullable, so a -1 placeholder is used
-            }
-        }
-    });
-
-    return (
-        <text x="47" y="60" className="fontBig">{ltsTest === 0 ? '88:88' : getDisplayString(elapsedTime, running)}</text>
-    );
+    if (seconds !== null) {
+        return `${Math.floor(Math.min(seconds, 5999) / 60).toString().padStart(2, '0')}${running ? ':' : ' '}${(Math.floor(Math.min(seconds, 5999) % 60)).toString().padStart(2, '0')}`;
+    }
+    return '';
 };
+
+export class Chrono extends DisplayComponent<ChronoProps> {
+    private readonly chronoText = Subject.create('');
+
+    private readonly elapsedTime = Subject.create(null);
+
+    private readonly running = Subject.create(false);
+
+    private readonly ltsTest = Subject.create(false);
+
+    private dcEssIsPowered: boolean;
+
+    private prevTime: number;
+
+    public onAfterRender(node: VNode): void {
+        super.onAfterRender(node);
+
+        const sub = this.props.bus.getSubscriber<ClockSimvars>();
+        sub.on('ltsTest').whenChanged().handle((ltsTest) => {
+            this.ltsTest.set(ltsTest === 0);
+        });
+
+        sub.on('dcHot1IsPowered').whenChanged().handle((dcHot1IsPowered) => {
+            if (!dcHot1IsPowered) {
+                this.running.set(false);
+                this.elapsedTime.set(null);
+            }
+        });
+
+        sub.on('dcEssIsPowered').whenChanged().handle((dcEssIsPowered) => this.dcEssIsPowered = dcEssIsPowered);
+
+        sub.on('absTime').atFrequency(5).handle((absTime) => {
+            if (this.running.get()) {
+                const newElapsedTime = (this.elapsedTime.get() || 0) + debouncedTimeDelta(absTime, this.prevTime);
+                this.elapsedTime.set(newElapsedTime);
+            }
+            this.prevTime = absTime;
+        });
+
+        const hEventsSub = this.props.bus.getSubscriber<HEvent>();
+        hEventsSub.on('hEvent').handle((eventName) => {
+            switch (eventName) {
+            case 'A32NX_CHRONO_RST':
+                if (this.dcEssIsPowered) {
+                    if (this.running.get()) {
+                        this.elapsedTime.set(0);
+                    } else {
+                        this.elapsedTime.set(null);
+                    }
+                }
+                break;
+            case 'A32NX_CHRONO_TOGGLE':
+                if (this.dcEssIsPowered) {
+                    this.running.set(!this.running.get());
+                }
+                break;
+            default: break;
+            }
+        });
+
+        this.elapsedTime.sub((elapsedTime) => SimVar.SetSimVarValue('L:A32NX_CHRONO_ELAPSED_TIME', 'number', elapsedTime ?? -1)); // Simvar ist not nullable, so a -1 placeholder is used
+
+        [
+            this.elapsedTime,
+            this.running,
+            this.ltsTest,
+        ].forEach((attr) => attr.sub(() => this.chronoText.set(getDisplayString(this.elapsedTime.get(), this.running.get(), this.ltsTest.get()))));
+    }
+
+    public render(): VNode {
+        return (
+            <text x="47" y="60" class="fontBig">{this.chronoText}</text>
+        );
+    }
+}
