+++ fbw-a32nx/src/systems/instruments/src/Clock/Components/ElapsedTime.tsx
@@ -1,39 +1,77 @@
-import React, { useEffect, useState } from 'react';
-import { useInteractionSimVar, useSimVar } from '@instruments/common/simVars';
-import { debouncedTimeDelta } from '../../Common';
+import { ComponentProps, DisplayComponent, EventBus, FSComponent, Subject, VNode } from 'msfssdk';
+import { debouncedTimeDelta } from '../shared/Utils';
+import { ClockSimvars } from '../shared/ClockSimvarPublisher';
 
-const getDisplayString = (seconds: number, running: boolean) : string => (seconds === 0 && !running ? ''
-    : `${Math.floor(Math.min(seconds, 359940) / 3600).toString().padStart(2, '0')}${running ? ':' : ' '}${(Math.floor((Math.min(seconds, 359940) % 3600) / 60)).toString().padStart(2, '0')}`);
+const getDisplayString = (seconds: number | null, running: boolean, ltsTest: boolean) : string => {
+    if (ltsTest) {
+        return '88:88';
+    }
 
-export const ElapsedTime = () => {
-    const [ltsTest] = useSimVar('L:A32NX_OVHD_INTLT_ANN', 'bool', 250);
-    const [dcEssIsPowered] = useSimVar('L:A32NX_ELEC_DC_ESS_BUS_IS_POWERED', 'bool', 250);
-    const [elapsedKnobPos] = useInteractionSimVar('L:A32NX_CHRONO_ET_SWITCH_POS', 'number', 'A32NX_CHRONO_ET_POS_CHANGED');
-    const [, setElapsedSimvarTime] = useSimVar('L:A32NX_CHRONO_ET_ELAPSED_TIME', 'number', 250);
-    const [absTime] = useSimVar('E:ABSOLUTE TIME', 'Seconds', 1000);
-    const [prevTime, setPrevTime] = useState(absTime);
-
-    const [elapsedTime, setElapsedTime] = useState(0);
-
-    useEffect(() => {
-        if (dcEssIsPowered) {
-            if (elapsedKnobPos === 0) {
-                const newElapsedTime = (elapsedTime || 0) + debouncedTimeDelta(absTime, prevTime);
-                setElapsedTime(newElapsedTime);
-                setElapsedSimvarTime(newElapsedTime);
-            } else if (elapsedKnobPos === 2) {
-                setElapsedTime(0);
-                setElapsedSimvarTime(-1); // Simvar is not nullable, so a -1 placeholder is used
-            }
-            setPrevTime(absTime);
-        }
-    }, [absTime, elapsedKnobPos, dcEssIsPowered]);
-
-    useEffect(() => {
-        setElapsedSimvarTime(-1);
-    }, []);
-
-    return (
-        <text x="47" y="247" className="fontBig">{ltsTest === 0 ? '88:88' : getDisplayString(elapsedTime, elapsedKnobPos === 0)}</text>
-    );
+    if (seconds !== null) {
+        return Math.floor(Math.min(seconds, 359940) / 3600).toString().padStart(2, '0')
+            + (running ? ':' : ' ')
+            + (Math.floor((Math.min(seconds, 359940) % 3600) / 60)).toString().padStart(2, '0');
+    }
+    return '';
 };
+
+interface ElapsedTimeProps extends ComponentProps {
+    bus: EventBus;
+}
+
+export class ElapsedTime extends DisplayComponent<ElapsedTimeProps> {
+    private readonly timerText = Subject.create('');
+
+    private readonly elapsedTime = Subject.create(null);
+
+    private readonly ltsTest = Subject.create(false);
+
+    private readonly elapsedKnobPos = Subject.create(1);
+
+    private dcEssIsPowered: boolean;
+
+    private prevTime: number;
+
+    onAfterRender(node: VNode): void {
+        super.onAfterRender(node);
+
+        const sub = this.props.bus.getSubscriber<ClockSimvars>();
+        sub.on('ltsTest').whenChanged().handle((ltsTest) => this.ltsTest.set(ltsTest === 0));
+
+        sub.on('dcHot1IsPowered').whenChanged().handle((dcHot1IsPowered) => {
+            if (!dcHot1IsPowered) {
+                this.elapsedTime.set(null);
+            }
+        });
+
+        sub.on('dcEssIsPowered').whenChanged().handle((dcEssIsPowered) => this.dcEssIsPowered = dcEssIsPowered);
+
+        sub.on('elapsedKnobPos').whenChanged().handle((elapsedKnobPos) => this.elapsedKnobPos.set(elapsedKnobPos));
+
+        sub.on('absTime').atFrequency(5).handle((absTime) => {
+            if (this.dcEssIsPowered) {
+                if (this.elapsedKnobPos.get() === 0) {
+                    const newElapsedTime = (this.elapsedTime.get() || 0) + debouncedTimeDelta(absTime, this.prevTime);
+                    this.elapsedTime.set(newElapsedTime);
+                } else if (this.elapsedKnobPos.get() === 2) {
+                    this.elapsedTime.set(null);
+                }
+                this.prevTime = absTime;
+            }
+        });
+
+        this.elapsedTime.sub((elapsedTime) => SimVar.SetSimVarValue('L:A32NX_CHRONO_ET_ELAPSED_TIME', 'number', elapsedTime ?? -1)); // Simvar ist not nullable, so a -1 placeholder is used
+
+        [
+            this.elapsedTime,
+            this.ltsTest,
+            this.elapsedKnobPos,
+        ].forEach((attr) => attr.sub(() => this.timerText.set(getDisplayString(this.elapsedTime.get(), this.elapsedKnobPos.get() === 0, this.ltsTest.get()))));
+    }
+
+    render(): VNode {
+        return (
+            <text x="47" y="247" class="fontBig">{this.timerText}</text>
+        );
+    }
+}
