--- fbw-a32nx/src/systems/instruments/src/Clock/config2.json
@@ -1,4 +1,0 @@
-{
-  "index": "./index.tsx",
-  "isInteractive": false
-}
