+++ fbw-a32nx/src/systems/instruments/src/Clock/shared/Utils.tsx
