--- fbw-a32nx/src/systems/instruments/src/Clock2/style.scss
@@ -1,48 +1,0 @@
-@font-face {
-  font-family: "AirbusChronometer";
-  src: url("/Fonts/AirbusChronometer.ttf") format("truetype");
-  font-weight: normal;
-  font-style: normal;
-}
-
-.off {
-  opacity: 0;
-  transition: opacity 0.01s linear;
-}
-
-.on {
-  opacity: 1;
-  transition: opacity 0.01s linear;
-}
-
-svg {
-  position: absolute;
-  width: 256px;
-  height: 256px;
-  background: rgba(5, 40, 80, 0.01);
-  /* TODO: Remove opacity when SAI/ISIS/CLOCK/BAT/ATC/RMP/RUD TRIM is split and not share the same emissive/mesh */
-  opacity: 0.5;
-}
-
-text {
-  font-family: "AirbusChronometer";
-  alignment-baseline: center;
-}
-
-.day text {
-  fill: rgba(220, 220, 220, 0.85);
-  transition: fill 5s linear;
-}
-
-.night text {
-  fill: rgba(220, 220, 220, 0.40);
-  transition: fill 5s linear;
-}
-
-.fontBig {
-  font-size: 65px;
-}
-
-.fontSmall {
-  font-size: 48px;
-}
